requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.4
langchain-community==0.0.16
pypdf2==3.0.1
youtube-transcript-api==0.6.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
alembic==1.13.1
httpx==0.26.0
